Change Dialog's `buttons` prop to an array of button descriptions
This is how Dialog should've been implemented. We pass an array of
objects that describe the buttons' behaviour. The most important thing
is that now onClose and onClick are different events, which had to be
handled only after closing the dialog. Now users see feedback faster, and
the code is much cleaner.